Add support for Game Pass version of Deep Rock
The Xbox Game Pass PC version of Deep Rock Galactic uses the
additional property "NameProperty". This property seems to only be used in
settings related information that is saved in the savefile.
This patch adds support for parsing such information, so that we do not
panic when encountering this property.